Add TimeDuration
